+++ hls4ml/converters/keras_v3/qkeras/batchnorm.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from ..core import KerasV3LayerHandler
-from .util import IsolatedLayerReader, set_default_config
+from .utils import IsolatedLayerReader, set_default_config
 
 
 class QKerasQConv2DBatchnormHandler(KerasV3LayerHandler):
