+++ hls4ml/converters/keras_v3/qkeras/utils.py
